Tune how the `default_settings.tbl` handles options it cannot find (#6656)
In-game options may be removed/inaccessible, such as VR mode removing the Window, but still valid for listing in the `default_settings.tbl` table.

In other words, we don't want error messages displaying if a player is playing  in VR mode (which will remove the Window option for that instance) but the window option is listed in the table.

Thus, downgrade the error to a debug print for options within the `default_settings.tbl`. As MJN stated, the table itself controls the parser here, so downgrading the warning is the best choice.