refine intrinsic_no_shields behavior
Delay processing of the `Intrinsic_no_shields` ship flag until the game is actually in-mission.  This avoids incorrectly setting ships in the loadout that may be changed from a shielded ship class to a non-shielded ship class, or vice versa.

Fixes buggy behavior reported by the Fate of the Galaxy mod.