Re-add setting eye orientation in `object_get_eye`
The refactor in 29804918dd8962be18b867d779e069f022cc6aa0 caused the GTB
Boanerges bomber to eh, look out of its ass, breaking the “Slaying
Ravana” mission and others. This is because `eye_orient` is no longer
set, and apparently most ships have a default orientation where this
is not noticed.

Re-add the assignment of the eye orientation.