switch sound effects to SDL3 device management
Continue using OpenAL for mixing, but feed the result to SDL3 for playback
rather than having OpenAL with the physical audio device. Outside of the minor
init changes OpenAL can be used as normal.

Bug fix: when querying the default playback device openal-soft will always
return "OpenAL Soft". So instead use the all_device_specifier if supported
(which it will be in this case) to get the proper default device name.