Fix AI Targeting of Subsystems
On current master, and ever since retail, if the AI is targeting a subsystem, and the target has shields lower than 0.25, then the AI aims straight for the subsystem, instead of leading the target. This is usually not that noticeable on retail ships where the larger ships with subsystems do not move quickly, but can become quite noticeable on mods where there are small fast shuttles and the AI becomes rather inept at disabling the engines.

Happy to tune variable and/or table entry names however.

Also took the time to cleanup an old define which had a not useful variable name and comment and was only used in one place not at all related to the comment.