reuse sexp audio handles
The `play-sound-from-file` sexp did not have any way to automatically recycle audio handles, instead relying on other sexps to explicitly close them.  It is more modder-friendly to implicitly free any handles of streams that have finished playing, so this PR adds that capability.  It also adds a Warning in the event that an unused audio handle cannot be found.

Fixes #7108.